Rename top page to mypage
